Change behavior when odd number of opt-ins
Depending on the context for the bot, the best behavior of how to deal
with an odd number of opt-ins is different. Before this commit, the
behavior was meant for a coffee-chat bot, where it makes sense to have
one group of three people for a coffee chat. In the context of a
pair-programming bot, it's better to only have pairs; to achieve that,
one person can have two pair-programming sessions. That's what this
commit does.